Dev/Contract 3 (Python Simple Storage)/requirements.txt
web3==5.23.1
py-solc-x